# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ambulanzsystem V1 API
  version: 0.1.0
  description: |
    Frozen Stage-0 contract for the Ambulanzsystem V1 (see docs/plan/master-plan.md).
    Locked decisions D1-D8 apply. JSON property names are serialized exactly as
    written here (no automatic camelCase transformation server-side).

    Auth token types (JWT `type` claim): `admin`, `leitstelle`, `user`, `qr`.
    Policies referenced below:
    - AdminOnly: type=admin
    - LeitstelleOrAdmin: type in (admin, leitstelle)
    - AuthenticatedUser: type in (admin, leitstelle, user)
    - TriageWrite: type in (admin, leitstelle, user, qr)

    Canonical triage values: rot | gelb | gruen | schwarz (ASCII, D5).
    `blau` and Unicode `grün` are rejected at the write boundary.
servers:
  - url: http://localhost:5000
    description: local backend
  - url: http://localhost:4010
    description: Prism mock server
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    post:
      tags: [users]
      summary: Create a user (admin, leitstelle, or responder). Policy AdminOnly. New admin/leitstelle users get requiresPasswordChange=true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string, maxLength: 255 }
                password: { type: string, minLength: 8 }
                role: { $ref: "#/components/schemas/Role" }
                accountType: { type: string, enum: [permanent, event], default: permanent }
                eventSceneId:
                  type: integer
                  description: Required when accountType=event; access is scoped to this top-level scene's subtree and its access window.
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Username already exists. }
  /api/users/change-password:
    post:
      tags: [users]
      summary: Change own password. Verifies current password, regenerates security stamp (invalidates all live tokens), revokes all refresh tokens. Policy AuthenticatedUser.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, newPassword]
              properties:
                username: { type: string }
                password: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "204": { description: Changed. All outstanding sessions for this user are now invalid. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}/revoke:
    post:
      tags: [users]
      summary: End a user's access early. Policy LeitstelleOrAdmin. Blocks new actions/reads; no forced client logout.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/self-cancel:
    post:
      tags: [users]
      summary: A responder/session ends its own access by choice. Policy TriageWrite.
      responses:
        "204": { description: Access ended. }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------- login QR codes
...
  /api/teams:
    post:
      tags: [teams]
      summary: Create a team in a scene. Policy LeitstelleOrAdmin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operationSceneId, name]
              properties:
                operationSceneId: { type: integer }
                name: { type: string, description: Name or radio call sign. }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    get:
      tags: [teams]
      summary: List teams for a scene. Policy TriageWrite.
      parameters:
        - name: operationSceneId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Teams.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/teams/{id}:
    put:
      tags: [teams]
      summary: Update any subset of optional coordination fields (FR-TEAM-07 — every field independently optional/clearable). Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: [string, "null"]
                  enum: [free, busy, unavailable, null]
                assignedPatientId: { type: [integer, "null"] }
                assignedLocation: { type: [string, "null"] }
                contactInfo: { type: [string, "null"] }
      responses:
        "200":
          description: Updated team.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ----------------------------------------------------------- persons
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: Missing/invalid/expired/revoked credentials. Expired access blocks new actions and reads (FR-AUTH-09).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but role not permitted (e.g. responder on admin endpoint, NFR-SEC-02).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [status, message]
      properties:
        status: { type: string, const: error }
        message: { type: string }
    Role:
      type: string
      enum: [admin, leitstelle, responder]
...
    User:
      type: object
      required: [id, username, role, accountType, createdAt, updatedAt]
      properties:
        id: { type: integer }
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        accountType: { type: string, enum: [permanent, event] }
        eventSceneId: { type: [integer, "null"] }
        requiresPasswordChange: { type: boolean }
        revokedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Team:
      type: object
      required: [id, operationSceneId, name, createdAt, updatedAt]
      properties:
        id: { type: integer }
        operationSceneId: { type: integer }
        name: { type: string }
        status:
          type: [string, "null"]
          enum: [free, busy, unavailable, null]
        assignedPatientId: { type: [integer, "null"] }
        assignedLocation: { type: [string, "null"] }
        contactInfo: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...